+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Revision history for Z-IO
 
+## 0.8.1.0  -- 2020-04-25
+
+* Add `getInterface` to `Z.IO.Network`.
+* `mkstemp` now return opend file, the type changed to `mkstemp :: CBytes -> Bool -> Resource (CBytes, File)`, which has an option for keep file or not.
+* `initTempFile` and `initTempDir` now do not need a prefix argument, the prefix is hardcoded as `Z-IO-`.
+
 ## 0.8.0.0  -- 2020-04-25
 
 This is an experimental version to test new 'BIO' module.